pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nexus-agent"
version = "0.1.0"
description = "让任何 AI 模型越用越好的开源 Agent 平台"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "typer>=0.12.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "rich>=13.0",
    "openai>=1.40.0",
    "anthropic>=0.34.0",
    "aiosqlite>=0.20.0",
]

[project.scripts]
nexus = "nexus.cli:main"

[project.optional-dependencies]
training = ["transformers>=4.45.0", "peft>=0.12.0", "torch>=2.0"]
